update: Change how layouts are fetched for more performant capital-P Page loads
